Combined Stm32u5xxPeripherals and Stm32u5xxDefaultPeripherals
