+++ chips/stm32u545/src/lib.rs
@@ -4,7 +4,7 @@
 
 #![no_std]
 
-pub use stm32u5xx::{chip, dma, exti, gpio, rcc, tim, usart, Stm32u5xxPeripherals};
+pub use stm32u5xx::{chip, dma, exti, gpio, rcc, tim, usart};
 
 use cortexm33::{CortexM33, CortexMVariant};
