+++ chips/stm32u5xx/src/lib.rs
@@ -49,32 +49,3 @@ pub unsafe fn init() {
     cortexm33::nvic::clear_all_pending();
     cortexm33::nvic::enable_all();
 }
-
-fn enable_tim2_clock() {
-    let rcc = rcc::Rcc::new(rcc::RCC_BASE);
-    rcc.enable_tim2();
-}
-
-pub struct Stm32u5xxPeripherals<'a> {
-    pub rcc: rcc::Rcc,
-    pub exti: &'a exti::Exti<'a>,
-    pub dma1: &'a dma::Dma,
-    pub gpio_a: gpio::Port<'a>,
-    pub gpio_c: gpio::Port<'a>,
-    pub usart1: &'a usart::Usart<'a>,
-    pub tim2: tim::Tim2<'a>,
-}
-
-impl<'a> Stm32u5xxPeripherals<'a> {
-    pub fn new(exti: &'a exti::Exti<'a>, dma1: &'a dma::Dma, usart1: &'a usart::Usart<'a>) -> Self {
-        Self {
-            rcc: rcc::Rcc::new(rcc::RCC_BASE),
-            exti,
-            dma1,
-            gpio_a: gpio::Port::new(gpio::GPIO_A_BASE, exti, gpio::GpioPort::PortA),
-            gpio_c: gpio::Port::new(gpio::GPIO_C_BASE, exti, gpio::GpioPort::PortC),
-            usart1,
-            tim2: tim::Tim2::new(tim::TIM2_BASE, enable_tim2_clock),
-        }
-    }
-}
